requirements.txt
boto3==1.34.147
botocore==1.34.147
jmespath==1.0.1
kafka-python==2.0.2
numpy==2.0.1
pandas==2.2.2
py4j==0.10.9.7
pyspark==3.5.1
python-dateutil==2.9.0.post0
pytz==2024.1
ruff==0.5.4
s3transfer==0.10.2
six==1.16.0
tzdata==2024.1
urllib3==2.2.2
